% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_varimp.R
\name{gg_varimp}
\alias{gg_varimp}
\title{Make A Ggplot Random Forest Variable Importance Plot}
\usage{
gg_varimp(random_forest_model, comparison)
}
\arguments{
\item{random_forest_model}{a random forest model object}

\item{comparison}{a string describing the model}
}
\description{
This function takes a random forest model objection, and a title description string, and produces a pretty ggplot variable imporance plot.
}
